Hooked up chat tabs to local chat and the debug log.
This is currently done in an ugly way. I wonder if it would be better to make the UIChatTab listen to the events from their respective message logs.

Also I question the need for keeping the text versions of the logs in the UIChatTab objects. Instead I feel like there should be "Actions" for loading and appending messages, operating on the assigned message log.

I don't see how to send generic parameters to a MonoBehaviour, so this became more tricky than anticipated.