Hooked up SendMessage to local chat.
This hook is ugly. We need a generic UIChatTab component in which we can specify what MessageLog to use, what function to call on SendMessage and what function to use for formatting incoming messages (i.e. convert to RTF). This should preferably be done using generics and action properties.

In addition, it would be nice if the "tab" object could be dragged off the main chat window and become the content of a separate floater and vice versa.