Improved resilience to scene switching.
The EventManager doesn't need to be a MonoBehaviour so it is now just a normal class.

For the other classes, I moved event registration from Start() to OnEnable() and unsubscribe in OnDisable(). This gets rid of straggling subscriptions.

It still won't maintain a session properly across scenes, but I don't think that we will have scene switching in the final product. If switching to the various test scene is done before logging in, it will be fine.