Build works again.
Since I use an assembly definition for the unit tests, it seems as if the "Editor" folder isn't enough to exclude the Editor folder scripts from builds. This could be remedied by adding am assembly definition in each Editor folder or, the way I did it, with a pre-processor directive in each editor script.